修复逻辑错误

+++ bot.py
@@ -275,15 +275,16 @@ def send_anime(message):
 @bot.message_handler(commands=['week'])
 def send_week(message):
     data = message.text.split(' ')
-    day = data[1]
-    week_data=week_text(day)
-    if data[0] == "/week" and day.isnumeric():
-            if 1<=int(day)<=7:
-                msg = bot.send_message(message.chat.id, "正在搜索请稍后...", reply_to_message_id=message.message_id, parse_mode='Markdown', timeout=20)
-                text = week_data['text']
-                markup = week_data['markup']
-                bot.delete_message(message.chat.id, message_id=msg.message_id, timeout=20)
-                bot.send_message(message.chat.id, text=text, parse_mode='Markdown', reply_markup=markup , timeout=20)
+    if len(data) == 2:
+        day = data[1]
+        if data[0] == "/week" and day.isnumeric():
+                if 1<=int(day)<=7:
+                    week_data=week_text(day)
+                    msg = bot.send_message(message.chat.id, "正在搜索请稍后...", reply_to_message_id=message.message_id, parse_mode='Markdown', timeout=20)
+                    text = week_data['text']
+                    markup = week_data['markup']
+                    bot.delete_message(message.chat.id, message_id=msg.message_id, timeout=20)
+                    bot.send_message(message.chat.id, text=text, parse_mode='Markdown', reply_markup=markup , timeout=20)
     else:
         bot.send_message(message.chat.id, "输入错误 请输入：`/week 1~7`", parse_mode='Markdown', timeout=20)
